FIX: Searches recursively for AppData folder
Now searches reclusively for AppData folder to delete database and modify web config

+++ tansferFiles.py
@@ -13,15 +13,22 @@ zippedSubmissionsFolder = './DDWT/Exam'
 #Remove the rtf coversheet files that you forgot to untick.
 removeRTFFiles = True
 #Limit the amount of files you want to move, good for testing. 0 = off
-limitFilesToMove = 1
+limitFilesToMove = 0
+
+'''
+
+		DDWT Options
+
+'''
 #Modify connection string for DDWT
 new_connection_string = "Data Source=(LocalDB)\MSSQLLocalDB;Initial Catalog=MoviesDB;Integrated Security=True"
+#Expected Folder Name
+expected_folder_name = "DDWTMovies"
 
 def create_folder_and_move_file(file_path, destination_folder):
 	# Extract the file name from the file path
 	file_name = os.path.basename(file_path)
 	extension = file_name[-4:] #get the last 4 digits of an extension
-	
 
 	if(extension.lower() == ".rtf" and removeRTFFiles == True):
 		os.remove(file_path)
@@ -41,18 +48,44 @@ def create_folder_and_move_file(file_path, destination_folder):
 	
 	print(f"Moved {file_name} to {destination_folder}")
 	DDWT_cleanup(destination_folder)
-	print(f"Cleaned up {destination_folder}")
 
 def DDWT_cleanup(folder_path):
-	AppData = folder_path + "/DDWTMovies/App_Data"
-	os.remove(AppData + "/MoviesDB.mdf")
-	os.remove(AppData + "/MoviesDB_log.ldf")
-	os.remove(AppData + "/SQLQuery MoviesDB.sql")
-	print(f"Deleted {AppData} + /MoviesDB.md")
+	attemptCount = 0
+	AppData = folder_path
+
+	logs = []
+	logs.append("Pre: {AppData}: AppData in folder?({not os.path.exists(AppData + '/App_Data')}), appData exists:{os.path.exists(AppData)}")
+
+	while(not os.path.exists(AppData + "/App_Data") and os.path.exists(AppData) and attemptCount < 3):
+		AppData += "/" + expected_folder_name
+		attemptCount+=1
+		logs.append(f"In Loop: {AppData}:{not os.path.exists(AppData + '/App_Data')}, appData exists:{os.path.exists(AppData)}")
+
+	logs.append(f"Post: {AppData}: AppData exists?{os.path.exists(AppData)}")
+
+	AppData += "/App_data"
+
+	if(not os.path.exists(AppData)):
+		print(f"ERROR: Unable to find App_Data folder for {folder_path}")
+		for log in logs:
+			print(f"{log}")
+		return
+	   
+	if(os.path.exists(AppData + "/MoviesDB.mdf")):
+		os.remove(AppData + "/MoviesDB.mdf")
+
+	if(os.path.exists((AppData + "/MoviesDB_log.ldf"))):
+		os.remove(AppData + "/MoviesDB_log.ldf")
+		
+	if(os.path.exists((AppData + "/SQLQuery MoviesDB.sql"))):
+		os.remove(AppData + "/SQLQuery MoviesDB.sql")
 
 	#Replace Connection String
-	WebConfigFile = folder_path + "/DDWTMovies/Web.config"
-	modify_connection_string(WebConfigFile, new_connection_string)
+	WebConfigFile = AppData + "/../Web.config"
+	if(os.path.exists(WebConfigFile)):
+		modify_connection_string(WebConfigFile, new_connection_string)
+	else:
+		print(f"UNABLE to clean {WebConfigFile}")
 
 def modify_connection_string(xml_file, new_connection_string):
 	tree = ET.parse(xml_file)
@@ -89,7 +122,7 @@ def bulk_move_files():
 	for file in files:
 		#Limit to a number of files to test
 		if(limitFilesToMove > 0 and counter > limitFilesToMove):
-			 continue
+			continue
 
 		#Get path to current file
 		pathToFile = path_to_file(file)
